requirements.txt
yt-dlp
openai
python-dotenv
streamlit
